Set up user Model
Set Up User Model
Laravel comes with a pre-defined User model; we can use the User model for authentication process. In this step, we will learn how to implement the jwt-auth package in a user model.

Define Tymon\JWTAuth\Contracts\JWTSubject contract before the User model. This method wants you to define the two methods:

getJWTIdentifier(): Get the identifier that will be stored in the subject claim of the JWT.
getJWTCustomClaims(): Return a key value array, containing any custom claims to be added to the JWT.
Open the app/Models/User.php file and replace the following code with the existing code.